Update run scripts to check for installed dependencies

+++ run.sh
@@ -3,6 +3,7 @@
 ORIG_DIR="$(pwd)"
 cd "$(dirname "$0")"
 BIN_DIR="$(pwd)"
+source functions
 
 export EXPOSE_PORT="${1:-8888}"
 
@@ -20,17 +21,22 @@ main() {
 }
 
 runWithDocker() {
-  docker run --rm --name jupyterlab \
-    --user root \
-    -p 8888:8888 \
-    -v ${PWD}/workshop:/home/${NB_USER}/work \
-    -e NB_USER="${NB_USER}" \
-    -e NB_UID="${NB_UID}" \
-    -e NB_GID="${NB_GID}" \
-    -e CHOWN_HOME="yes" \
-    -e JUPYTER_ENABLE_LAB="yes" \
-    eoepca/workshop \
-    start-notebook.sh --NotebookApp.token=\'\'
+  if hash docker 2>/dev/null; then
+    docker run --rm --name jupyterlab \
+      --user root \
+      -p 8888:8888 \
+      -v ${PWD}/workshop:/home/${NB_USER}/work \
+      -e NB_USER="${NB_USER}" \
+      -e NB_UID="${NB_UID}" \
+      -e NB_GID="${NB_GID}" \
+      -e CHOWN_HOME="yes" \
+      -e JUPYTER_ENABLE_LAB="yes" \
+      eoepca/workshop \
+      start-notebook.sh --NotebookApp.token=\'\'
+  else
+    error "ERROR: docker must be installed - not found. Aborting..."
+    return 1
+  fi
 }
 
 main "$@"
